switch package manager to poetry and rename package from metazerse to offchain
