--- requirements.txt
@@ -1,6 +1,0 @@
-pre-commit==2.15.0
-pydantic==1.9.0
-pytest==6.2.5
-python-json-logger==2.0.2
-requests==2.26.0
-urllib3==1.26.7
